moved plone specific stuff into a new repos

+++ addCustomBuildoutLayers.sh
@@ -3,13 +3,23 @@
 DEST=`python -c "import os;s=os.path.expanduser('~');print s"`
 DEST=$DEST/Plone4
 
-echo 'rename the original buildout.cfg'
-mv $DEST/zeocluster/buildout.cfg $DEST/zeocluster/unified-buildout.cfg
 
-echo 'add the new buildout layers'
-cp *.cfg $DEST/zeocluster
+if [ -f "${DEST}/unified-buildout.cfg" ]
+then
+  echo "skipping install of buildout layers"
+else
 
-echo 're-run buildout to add the new buildout layers'
-cd $DEST/zeocluster
-./bin/buildout -v 
+  echo 'rename the original buildout.cfg'
+  mv $DEST/zeocluster/buildout.cfg $DEST/zeocluster/unified-buildout.cfg
+
+  echo 'add the new buildout layers'
+  cp *.cfg $DEST/zeocluster
+
+  chown plone:plone *.cfg
+
+  echo 're-run buildout to add the new buildout layers'
+  cd $DEST/zeocluster
+  ./bin/buildouti
+
+fi
